moved function create_meta_info to app/functions to ease future testing
