+++ app/routes/page.edit.route.js
@@ -2,8 +2,9 @@
 'use strict';
 
 // Modules
-var fs           = require('fs');
-var get_filepath = require('../functions/get_filepath.js');
+var fs               = require('fs');
+var get_filepath     = require('../functions/get_filepath.js');
+var create_meta_info = require('../functions/create_meta_info.js');
 
 function route_page_edit (config, raneto) {
   return function (req, res, next) {
@@ -35,28 +36,6 @@ function route_page_edit (config, raneto) {
 
     // Create content including meta information (i.e. title, description, sort)
     function create_content(body) {
-
-      function create_meta_info(meta_title, meta_description, meta_sort) {
-        var newline = '\n';
-        var comment_start = '/*';
-        var comment_end = '*/';
-        var title_label = 'Title: ';
-        var description_label = 'Description: ';
-        var description_sort = 'Sort: ';
-
-        var no_meta_info_is_present = !meta_title && !meta_description && !meta_sort;
-        var meta_header = '';
-        if (!no_meta_info_is_present) {
-          meta_header = comment_start + newline +
-            (meta_title ? title_label + meta_title + newline : '') +
-            (meta_description ? description_label + meta_description + newline : '') +
-            (meta_sort ? description_sort + meta_sort + newline : '') +
-            comment_end + newline + newline;
-        }
-
-        return meta_header;
-      }
-    
       var meta = create_meta_info(body.meta_title, body.meta_description, body.meta_sort);
       return meta + body.content;
     }
